# docs/swagger.yaml
basePath: /
definitions:
  base.BaseResponse:
    description: BaseResponse is the base response for all the endpoints
    properties:
      data: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
  request.CreateUserUsageRequest:
    description: CreateUserUsageRequest is the request for the create user-usage endpoint
    properties:
      date:
        type: string
    type: object
...
  request.LoginRequest:
    description: LoginRequest is the request for the login endpoint
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  request.RegisterRequest:
    description: RegisterRequest is the request for the register endpoint
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  response.AuthResponse:
    description: AuthResponse is the response for the auth controller
    properties:
      email:
        type: string
      id:
        type: integer
      token:
        type: string
    type: object
...
  response.UserUsageResponse:
    description: UserUsageResponse is the response for the user usage endpoint
    properties:
      date:
        type: string
      id:
        type: integer
      total_cost:
        type: number
      total_energy:
        type: number
      user_id:
        type: integer
    type: object
...
host: http://52.65.161.24:8000
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple API for managing energy usage in a household.
  termsOfService: http://swagger.io/terms/
  title: Energia API Mini Project
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Log in a user using email and password
      parameters:
      - description: Login Request Body
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuthResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/base.BaseResponse'
      summary: User Login
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register Request Body
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuthResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/base.BaseResponse'
      summary: User Registration
      tags:
      - Auth
...
  /user-usage:
    get:
      consumes:
      - application/json
      description: Find user usage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserUsageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/base.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Find user usage
      tags:
      - User Usage
    post:
      consumes:
      - application/json
      description: Create a new user usage
      parameters:
      - description: User usage to create
        in: body
        name: user_usage
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserUsageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserUsageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/base.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new user usage
      tags:
      - User Usage
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"